added delete feature

+++ src/components/DeleteEmpButton.js
@@ -8,19 +8,27 @@ function DeleteEmpButton(emp) {
     <button
       className="btn btn-danger centered "
       onClick={() => {
-        alert(
-            "YOU ARE ABOUT TO DELETE THE FOLLOWING EMPLOYEE :" +
+      if(window.confirm(
+            "YOU ARE ABOUT TO DELETE THE FOLLOWING EMPLOYEE :\r\n" +
               emp.emp.firstName +
-              ", " +
+              " " +
               emp.emp.lastName +
-              ", " +
-              emp.emp.email+"! THIS CANNOT BE UNDONE! PRESS OK TO PROCEED OR CANCEL TO RETURN TO EMPLOYEE LIST"
-          );
+              ", " +     
+              emp.emp.email+"\r\n THIS CANNOT BE UNDONE!\r\n PRESS OK TO PROCEED OR CANCEL TO RETURN TO EMPLOYEE LIST"
 
-          EmployeeService.deleteEmployee(emp.emp.id);
+              
+          ) == true){
+            EmployeeService.deleteEmployee(emp.emp.id);
+            alert("Employee deleted!");
+            navigate(`/employees` );
 
-          alert("Employee deleted!")
-        navigate(`/employees` );
+          } else{
+            navigate(`/employees`)
+          };
+
+          
+
+          
       }}
     >
       Delete
